python code/requirements.txt
pyserial
sqlite3